Moved files to rascsi/rasctl folders
