+++ src/raspberrypi/rascsi.cpp
@@ -23,11 +23,11 @@
 #include "protobuf_serializer.h"
 #include "command_util.h"
 #include "rascsi_version.h"
-#include "rascsi_executor.h"
-#include "rascsi_response.h"
-#include "rascsi_image.h"
 #include "rascsi_interface.pb.h"
-#include "rascsi_service.h"
+#include "rascsi/executor.h"
+#include "rascsi/response.h"
+#include "rascsi/image.h"
+#include "rascsi/service.h"
 #include "rasutil.h"
 #include "spdlog/spdlog.h"
 #include "spdlog/sinks/stdout_color_sinks.h"
