+++ src/raspberrypi/rascsi/executor.cpp
@@ -14,15 +14,15 @@
 #include "devices/primary_device.h"
 #include "devices/disk.h"
 #include "devices/file_support.h"
-#include "rascsi_service.h"
-#include "rascsi_response.h"
-#include "rascsi_image.h"
+#include "service.h"
+#include "response.h"
+#include "image.h"
 #include "rascsi_exceptions.h"
 #include "localizer.h"
 #include "command_util.h"
 #include "rasutil.h"
 #include "spdlog/spdlog.h"
-#include "rascsi_executor.h"
+#include "executor.h"
 #include <sstream>
 
 using namespace spdlog;
